+++ src/ts/drawutilssvg.ts
@@ -34,7 +34,8 @@
  * @modified 2022-03-26 Added the private `nodeDefs` and `bufferedNodeDefs` attributes.
  * @modified 2022-03-26 Added the `texturedPoly` function to draw textures polygons.
  * @modified 2022-07-26 Adding `alpha` to the `image(...)` function.
- * @version  1.6.1
+ * @modified 2022-11-10 Tweaking some type issues.
+ * @version  1.6.2
  **/
 
 import { CircleSector } from "./CircleSector";
@@ -1321,10 +1322,10 @@ export class drawutilssvg implements DrawLib<void | SVGElement> {
     this.curId = curId + "_text";
     const textNode: SVGElement = this.makeNode("text");
     node.appendChild(textNode);
-    textNode.setAttribute("font-family", options.fontFamily); // May be undefined
-    textNode.setAttribute("font-size", options.fontSize ? `${options.fontSize * this.scale.x}` : null);
-    textNode.setAttribute("font-style", options.fontStyle ? `${options.fontStyle}` : null);
-    textNode.setAttribute("font-weight", options.fontWeight ? `${options.fontWeight}` : null);
+    textNode.setAttribute("font-family", options.fontFamily ?? ""); // May be undefined
+    textNode.setAttribute("font-size", options.fontSize ? `${options.fontSize * this.scale.x}` : "");
+    textNode.setAttribute("font-style", options.fontStyle ? `${options.fontStyle}` : "");
+    textNode.setAttribute("font-weight", options.fontWeight ? `${options.fontWeight}` : "");
     textNode.setAttribute("text-anchor", textAlign);
     textNode.setAttribute("transform-origin", "0 0");
     textNode.setAttribute("transform", rotate);
@@ -1488,19 +1489,7 @@ export class drawutilssvg implements DrawLib<void | SVGElement> {
     const _sy = (index: number): void => {
       data[index] = scale.y * Number(data[index]);
     };
-    const stx = (value: number): number => {
-      return offset.x + scale.x * value;
-    };
-    const sty = (value: number): number => {
-      return offset.y + scale.y * value;
-    };
-    // scale only {x,y}
-    const sx = (value: number): number => {
-      return scale.x * value;
-    };
-    const sy = (value: number): number => {
-      return scale.y * value;
-    };
+
     var i: number = 0;
     var lastPoint: XYCoords = { x: NaN, y: NaN };
     // "save last point"