+++ changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
-- 2022-11-06
+- 2022-11-10
+  - Fixing an export bug: VEllipseSector was missing in the CJS and ESM exports.
+- 2022-11-06 (v1.19.0)
   - Adding an XML declaration to the SVG export routine.
   - Added the CairoTileBuilder to the Truchet demo 44.
 - 2022-10-25